// src/app/models/superApp/IUser.ts
export interface IUser {
    custId: number;
    firstName: string;
    lastName: string;
    walletAcctNumber: string;
    dob: string;
    gender: string;
    phoneNumber: string;
    emailAddress: string;
    contactAddress: string;
    bvn: string;
    Token: string;
    ProfilePicPath: string;
    nokFullName: string;
    nokRelationship: string;
    nokEmailAddress: string;
    nokPhone: string;
    isBvnSET: boolean;
    isFidelityCustomerValidated: boolean;
    finaclePhoneNumber: string;
}